+++ app/src/main/java/com/vertice/teepop/leaveapp/util/bus/ApproveMessageEvent.kt
@@ -4,6 +4,6 @@ package com.vertice.teepop.leaveapp.util.bus
  * Created by VerDev06 on 1/9/2018.
  */
 data class ApproveMessageEvent(var comment: String = "",
-                               var leaveId: Int = 0) {
+                               var leaveId: Long = 0) {
 
 }
